+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # What's New with Enterprise-NG
 
+## v4.13.0
+
+### 4.13.0 Features
+
+### 4.13.0 Fixes
+
+### 4.13.0 Chore & Maintenance
+
 ## v4.12.0
 
 ### 4.12.0 Features
@@ -18,7 +26,7 @@
 
 ### 4.12.0 Chore & Maintenance
 
-- `[General]` `d3` and `@types/d3` version 4.13 added as a dependency for the `ids-enterprise-ng` package.  
+- `[General]` `d3` and `@types/d3` version 4.13 added as a dependency for the `ids-enterprise-ng` package.
     Developers should change their angular.json files to refer to the installed version of d3 (and their types) and not the copy in the `enterprise` folder.
 
 ## v4.11.0